refactor(repository): update database and repository layer
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)

+++ lib/repository/repository_impl.dart
@@ -1,5 +1,6 @@
 import 'package:app/model/write_record.dart';
 import 'package:app/repository/repository.dart';
+import 'package:app/utils/platform_utils.dart';
 import 'package:sqflite/sqflite.dart';
 
 class RepositoryImpl implements Repository {
@@ -12,20 +13,37 @@ class RepositoryImpl implements Repository {
   @override
   Stream<Iterable<WriteRecord>> subscribeWriteRecordList() {
     return _subscriptionManager.createStream(() async {
-      return _db.query('record')
-        .then((value) => value.map((e) => WriteRecord.fromJson(e)));
+      return _db
+          .query('record')
+          .then((value) => value.map((e) => WriteRecord.fromJson(e)));
     });
   }
 
   @override
   Future<WriteRecord> createOrUpdateWriteRecord(WriteRecord record) async {
-    final id = await _db.insert('record', record.toJson(), conflictAlgorithm: ConflictAlgorithm.replace);
+    // On HarmonyOS, sqflite operations may have limitations
+    if (PlatformUtils.isOHOS) {
+      print('Create/Update operation not supported on HarmonyOS yet');
+      return record;
+    }
+    if (record.id == null) {
+      final id = await _db.insert('record', record.toJson());
+      _subscriptionManager.publish();
+      return record.copyWith(id: id);
+    }
+    await _db.update('record', record.toJson(),
+        where: 'id = ?', whereArgs: [record.id!]);
     _subscriptionManager.publish();
-    return WriteRecord.fromJson(record.toJson()..[WriteRecord.ID] = id); // todo: copyWith
+    return record;
   }
 
   @override
   Future<void> deleteWriteRecord(WriteRecord record) async {
+    // On HarmonyOS, sqflite operations may have limitations
+    if (PlatformUtils.isOHOS) {
+      print('Delete operation not supported on HarmonyOS yet');
+      return;
+    }
     await _db.delete('record', where: 'id = ?', whereArgs: [record.id!]);
     _subscriptionManager.publish();
   }
